% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSI_Lab_2_package_ans.R
\name{statistics_of_drg_code}
\alias{statistics_of_drg_code}
\title{Statistics for all of the DRG codes for average Medicare payments}
\usage{
statistics_of_drg_code(data, statistics)
}
\arguments{
\item{data}{a dataframe}

\item{statistics}{a string name for the calculation}
}
\value{
Mean, median or standard deviation for average medicare payments
}
\description{
This function calculates the mean, median, standard deviation.
}
\examples{
library(readr)
drg_data <- read_csv("DRG_data.csv")

statistics_of_drg_code(drg_data, "mean")

}
